requirements.txt
attrs==22.1.0
ConfigParser==6.0.0
contextlib2==21.6.0
cryptography==41.0.2
Cython==3.0.2
dl==0.1.0
docutils==0.18.1
docutils==0.18.1
HTMLParser==0.0.2
ipython==8.12.0
ipywidgets==8.0.4
Jinja2==3.1.2
jnius==1.1.0
keyring==23.13.1
lockfile==0.12.2
mock==5.1.0
numpy==1.24.3
pandas==1.5.3
Pillow==9.4.0
Pillow==10.0.1
protobuf==4.24.3
pyOpenSSL==23.2.0
pyOpenSSL==23.2.0
railroad==0.5.0
Sphinx==5.0.2
tornado==6.3.2
trove_classifiers==2023.9.19
truststore==0.8.0
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
yahoofinancials==1.16
yfinance==0.2.30
google-cloud-storage
google-api-python-client